+++ integration-tests/__tests__/require_after_teardown.test.js
@@ -1,0 +1,23 @@
+/**
+ * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
+'use strict';
+
+const runJest = require('../runJest');
+
+test('prints useful error for requires after test is done', () => {
+  const {stderr} = runJest('require-after-teardown');
+
+  const interestingLines = stderr
+    .split('\n')
+    .slice(9, 20)
+    .join('\n');
+
+  expect(interestingLines).toMatchSnapshot();
+});