+++ packages/jest-runtime/src/index.js
@@ -19,6 +19,7 @@ import type {SourceMapRegistry} from 'types/SourceMaps';
 
 import path from 'path';
 import HasteMap from 'jest-haste-map';
+import {formatStackTrace, separateMessageFromStack} from 'jest-message-util';
 import Resolver from 'jest-resolve';
 import {createDirectory, deepCyclicCopy} from 'jest-util';
 import {escapePathForRegex} from 'jest-regex-util';
@@ -551,9 +552,24 @@ class Runtime {
       }
     }
 
-    const wrapper = this._environment.runScript(transformedFile.script)[
-      ScriptTransformer.EVAL_RESULT_VARIABLE
-    ];
+    const runScript = this._environment.runScript(transformedFile.script);
+
+    if (runScript === null) {
+      const {message, stack} = separateMessageFromStack(
+        new ReferenceError(
+          'You are trying to `import` a file after the Jest environment has been torn down.',
+        ).stack,
+      );
+
+      console.error(
+        `\n${message}\n` +
+          formatStackTrace(stack, this._config, {noStackTrace: false}),
+      );
+      process.exitCode = 1;
+      return;
+    }
+
+    const wrapper = runScript[ScriptTransformer.EVAL_RESULT_VARIABLE];
     wrapper.call(
       localModule.exports, // module context
       localModule, // module object